:art: Improving structure of the code
